[refactor] table preprocessing: big code refactor Part 2
- refactor c74-independant code into SamplePreprocessor, separate responsibilities
- add dash pattern to painter interface
