+++ source/projects/bfa.table_preprocessing/bfa.table_preprocessing.cpp
@@ -4,16 +4,9 @@
 /// @license    Use of this source code is governed by the MIT License found in the License.md file.
 
 #include "c74_min.h"
-
-#include "waveform_processing.h"
-#include "pitch_detection.h"
 #include "sample_preprocessor.h"
-#include "wavetable_oscillator.h"
-#include "graphics_transform.h"
-#include "min_event_wrapper.h"
-#include "min_painter.h"
+#include "sample_preprocessor.cpp" // cannot include c74_min.h in more than one translation unit
 
-#include <chrono>
 
 using namespace c74::min;
 using namespace c74::min::ui;
@@ -40,24 +33,19 @@ void drawPoint(target& t, const Point& p1, const color& c, double siz = 1.0) {
 
 
 
-class table_preprocessing : public object<table_preprocessing>, public ui_operator<160, 80>
+class table_preprocessing : public object<table_preprocessing>, public ui_operator<160, 80>, public Callback
 {
 public:
 	using Wavetable = Wavetable<float>;
 	using Osc = WavetableOscillator<Wavetable>;
 
-	enum class Mode {
-		free,
-		zeros,
-		period
-	};
 
 	MIN_DESCRIPTION{ "Read from a buffer~ and display." };
 	MIN_TAGS{ "audio, sampling, ui, time" };
 	MIN_AUTHOR{ "BFA_JK" };
 	MIN_RELATED{ "index~, buffer~, wave~, waveform~" };
 
-	table_preprocessing(const atoms& args = {}) : ui_operator::ui_operator{ this, args } {}
+	table_preprocessing(const atoms& args = {}) : ui_operator::ui_operator{ this, args }, samplePreprocessor(*this) {}
 
 	~table_preprocessing() = default;
 
@@ -101,7 +89,7 @@ public:
 
 	message<> dspsetup{ this, "dspsetup",
 		[this](const c74::min::atoms& args, const int inlet) -> c74::min::atoms {
-			setup(static_cast<float>(args[0]));
+			samplePreprocessor.setup(static_cast<float>(args[0]));
 			return {};
 		} };
 
@@ -109,12 +97,13 @@ public:
 	message<> setMode{
 		this, "setMode", "Set frame selection mode (custom, period).", [this](const c74::min::atoms& args, const int inlet) -> c74::min::atoms {
 			if (args[0] == "Free") {
-				setModeImpl(Mode::free);
+				samplePreprocessor.setModeImpl(SamplePreprocessor::Mode::free);
 			} else if (args[0] == "Zeros") {
-				setModeImpl(Mode::zeros);
+				samplePreprocessor.setModeImpl(SamplePreprocessor::Mode::zeros);
 			} else if (args[0] == "Period") {
-				setModeImpl(Mode::period);
+				samplePreprocessor.setModeImpl(SamplePreprocessor::Mode::period);
 			}
+			redraw();
 			return {};
 		}
 	};
@@ -129,7 +118,7 @@ public:
 	message<> mousewheel{
 		this, "mousewheel", [this](const c74::min::atoms& args, const int inlet) -> c74::min::atoms {
 			event e{ args };
-			mousewheelImpl(createMouseEvent(e, MouseEvent::Action::Wheel));
+			samplePreprocessor.mousewheelImpl(createMouseEvent(e, MouseEvent::Action::Wheel));
 			return {};
 		}
 	};
@@ -138,7 +127,7 @@ public:
 	message<> mousedown{
 		this, "mousedown", [this](const c74::min::atoms& args, const int inlet) -> c74::min::atoms {
 			event e{ args };
-			mousedownImpl(createMouseEvent(e, MouseEvent::Action::Down));
+			samplePreprocessor.mousedownImpl(createMouseEvent(e, MouseEvent::Action::Down));
 			return {};
 		}
 	};
@@ -146,15 +135,7 @@ public:
 	message<> mouseup{
 		this, "mouseup", [this](const c74::min::atoms& args, const int inlet) -> c74::min::atoms {
 			event e{ args };
-			mouseupImpl(createMouseEvent(e, MouseEvent::Action::Up));
-			return {};
-		}
-	};
-
-	message<> dblclick{
-		this, "dblclick", [this](const c74::min::atoms& args, const int inlet) -> c74::min::atoms {
-			event e{ args };
-			resetTransform();
+			samplePreprocessor.mouseupImpl(createMouseEvent(e, MouseEvent::Action::Up));
 			return {};
 		}
 	};
@@ -162,11 +143,10 @@ public:
 	message<> mousedrag{
 		this, "mousedrag", [this](const c74::min::atoms& args, const int inlet) -> c74::min::atoms {
 			event e{ args };
-			mousedragImpl(createMouseEvent(e, MouseEvent::Action::Drag));
+			samplePreprocessor.mousedragImpl(createMouseEvent(e, MouseEvent::Action::Drag));
 			return {};
 		}
 	};
-
 
 
 	message<> generate_frame{
@@ -179,76 +159,23 @@ public:
 	message<> paint{
 		this, "paint", [this](const c74::min::atoms& args, const int inlet) -> c74::min::atoms {
 			target t{ args };
-			if (targetSize.x != t.width() || targetSize.y != t.height()) {
-				targetResized(t.width(), t.height());
-			}
 
 			rect<fill> rect{ t, color{ backgroundColor } };
 			MaxPainter painter{ t };
+			samplePreprocessor.draw(painter, { to(waveformColor), to(zeroCrossingsColor), to(overlayColor), to(draggingRectColor), strokeWidth });
 
-			if (samplePreprocessor.inputSamples.size() > 0) {
-				painter.fillColor(to(overlayColor));
-				drawOverlayRects(painter);
-
-				if ((samplePreprocessor.zeroCrossings.size() > 0) && (mode == Mode::zeros)) {
-					painter.strokeColor(to(zeroCrossingsColor));
-					painter.strokeWidth(strokeWidth);
-					drawZeroCrossings(painter);
-				}
-
-				painter.strokeWidth(strokeWidth);
-				painter.strokeColor(to(waveformColor));
-				painter.fillColor(to(waveformColor));
-				drawSamples(painter);
-
-				painter.strokeColor(to(draggingRectColor));
-				drawDraggingRect(painter);
-			}
 
 			return {};
 		}
 	};
 
-	void setup(float sampleRate);
-	void setModeImpl(Mode newMode);
+	void doRedraw() override {
+		redraw();
+	}
 
-	void targetResized(double width, double height); // Called when the target has been resized through any means
-	void constrainViewTransform();
-	void drawSamples(MaxPainter& painter);
-	void drawDraggingRect(MaxPainter& painter);
-	void drawZeroCrossings(MaxPainter& painter);
-	void drawOverlayRects(MaxPainter& painter);
-
-	void inputSamplesChanged();
-	void resetTransform();
-	void sampleDroppedImpl();
-	void setSampleData(const std::vector<float>& data);
-	bool canExport() const;
-	std::pair<int, int> getCurrentExportRange() const;
-	bool exportFrame();
-	void notifyCanExportStatus();
-
-	void updateFreeSelection(const Point& mouseDownPoint, const Point& currentMousePoint);
-	void updateZerosSelection(const Point& mouseDownPoint, const Point& currentMousePoint);
-
-
-	void mousedownImpl(const MouseEvent& e);
-	void mousedragImpl(const MouseEvent& e);
-	void mouseupImpl(const MouseEvent& e);
-	void mousewheelImpl(const MouseEvent& e);
-
-	// Helper functions
-	double nearestZeroCrossing(double sampleIdx) const;
-	MouseEvent::Button getButton(const event& e) const;
-
-
-	// Getters for testing
-	Mode getMode() const { return mode; }
-	bool isDragging() const { return dragging; }
-	double getSampleRate() const { return sampleRate; }
-	Rect getDataRange() const { return dataRange; }
-	std::pair<double, double> getFreeSelection() const { return freeSelection; }
-
+	void doNotifyCanExportStatus() override {
+		notifyCanExportStatus();
+	}
 	//    void analyse_period_zero_crossings()
 	//    {
 	//
@@ -279,96 +206,31 @@ public:
 	//}
 
 	//private:
-
-	float margin{ 10.f }; //Könnte auch als Attribut
-	float sampleRate{ 48000.f };
-	Mode mode{ Mode::free };
+	void sampleDroppedImpl();
+	void setSampleData(const std::vector<float>& data);
+	void notifyCanExportStatus();
+	bool exportFrame();
+	MouseEvent::Button getButton(const event& e) const;
 
 	SamplePreprocessor samplePreprocessor;
-	// std::optional<PitchInfo> pitchInfoOptional;
-	// PitchInfo pitchInfo;
-	// float tf0, deviationInSmps{10.f};
-
-
-	// mouse event related
-	Point mouseDownPoint, currentMousePoint;
-	MouseEvent::Button button{};
-	bool dragging{ false };
-	std::chrono::time_point<std::chrono::system_clock> clickTime;
-
-	// transform related
-	Rect dataRange;						// size of the draw target, will be updated on first draw.
-	Point targetSize{ 100, 100 };		// size of the draw target, will be updated on first draw.
-	Rect waveformView{ {}, targetSize }; // size of the draw target, will be updated on first draw.
-	Transform transform;
-	std::pair<double, double> freeSelection;
-	std::pair<double, double> zerosSelection;
 };
 
 
-void table_preprocessing::setup(float sampleRate) {
-	this->sampleRate = sampleRate;
-	cout << "dspsetup happend" << endl;
-}
+MIN_EXTERNAL(table_preprocessing);
 
-void table_preprocessing::setModeImpl(Mode newMode) {
-	mode = newMode;
-	redraw();
-	notifyCanExportStatus();
-}
-
-
-void table_preprocessing::updateFreeSelection(const Point& mouseDownPoint, const Point& currentMousePoint) {
-	const auto selectionValueRect = transform.from({ mouseDownPoint, currentMousePoint });
-	freeSelection.first = std::clamp(selectionValueRect.x, 0., samplePreprocessor.inputSamples.size() - 1.0);
-	freeSelection.second = std::clamp(selectionValueRect.x + selectionValueRect.width, 0., samplePreprocessor.inputSamples.size() - 1.0);
-	notifyCanExportStatus();
-}
-
-void table_preprocessing::updateZerosSelection(const Point& mouseDownPoint, const Point& currentMousePoint) {
-	double nearestCrossing1 = nearestZeroCrossing(transform.fromX(mouseDownPoint.x));
-	double nearestCrossing2 = nearestZeroCrossing(transform.fromX(currentMousePoint.x));
-	zerosSelection.first = std::clamp(nearestCrossing1, 0., samplePreprocessor.inputSamples.size() - 1.0);
-	zerosSelection.second = std::clamp(nearestCrossing2, 0., samplePreprocessor.inputSamples.size() - 1.0);
-	if (zerosSelection.second < zerosSelection.first) { std::swap(zerosSelection.second, zerosSelection.first); }
-	notifyCanExportStatus();
-}
-
-bool table_preprocessing::canExport() const {
-	const auto [begin, end] = getCurrentExportRange();
-	if (samplePreprocessor.inputSamples.empty()) return false;
-	if (end <= begin) return false;
-	if (begin < 0 || end >= samplePreprocessor.inputSamples.size()) return false;
-	if (std::abs(begin - end) < Wavetable::minimumInputSize()) return false;
-	return true;
-}
-
-std::pair<int, int> table_preprocessing::getCurrentExportRange() const {
-	if (mode == Mode::free) {
-		return { std::round(freeSelection.first), std::round(freeSelection.second) };
-	} else if (mode == Mode::zeros) {
-		return { std::round(zerosSelection.first), std::round(zerosSelection.second) };
-	}
-	return {};
-}
 
 bool table_preprocessing::exportFrame() {
-	if (!canExport()) return false;
-	const auto [begin, end] = getCurrentExportRange();
-
 	targetBuffer.set(targetBufferName);
 	buffer_lock<false> buf(targetBuffer);
 	if (!buf.valid()) return false;
-	const int targetTablesize = buf.frame_count();
-	const std::vector<float> selectedSamples{ samplePreprocessor.inputSamples.begin() + begin, samplePreprocessor.inputSamples.begin() + end };
-	const float exportTableOscFreq = sampleRate / static_cast<float>(targetTablesize);
+	const auto size = buf.frame_count();
 
-	const Wavetable table{ selectedSamples, sampleRate / 2.f };
-	std::vector<Wavetable> wavetable{ table };
-	Osc interpolationOscillator{ &wavetable, sampleRate, exportTableOscFreq };
-
-	for (int i = 0; i < targetTablesize; i++) {
-		buf[i] = interpolationOscillator++;
+	auto result = samplePreprocessor.exportFrame(size);
+	if (result) {
+		auto data = *result;
+		for (int i = 0; i < size; i++) {
+			buf[i] = data[i];
+		}
 	}
 	outletStatus.send("newFrame");
 	buf.dirty();
@@ -376,126 +238,12 @@ bool table_preprocessing::exportFrame() {
 }
 
 void table_preprocessing::notifyCanExportStatus() {
-	if (canExport()) {
+	if (samplePreprocessor.canExport()) {
 		outletStatus.send("CanExportStatus", 1);
 	} else {
 		outletStatus.send("CanExportStatus", 0);
 	}
 }
-
-void table_preprocessing::drawSamples(MaxPainter& painter) { //Das allgemeingültig schreiben und auch bei StackedFrames verwenden!
-	if (samplePreprocessor.inputSamples.empty()) return;
-
-	// get first/last visible sample
-	const int first = std::max(0., transform.fromX(0) - 1.);
-	const int last = std::min<double>(samplePreprocessor.inputSamples.size(), std::ceil(transform.fromX(targetSize.x) + 1.));
-	const int step = std::max<int>(1, (last - first) / (targetSize.x * 10.));
-
-	if (step == 1) {
-		const auto point = transform.apply({ static_cast<double>(first), -samplePreprocessor.inputSamples[first] * waveformYScaling });
-		auto previous = point;
-		for (int i = first + 1; i < last; ++i) {
-			const auto point = transform.apply({ static_cast<double>(i), -samplePreprocessor.inputSamples[i] * waveformYScaling });
-			painter.line(previous, point);
-			previous = point;
-		}
-	} else if (step < 10) {
-		const auto point = transform.apply({ static_cast<double>(first), -samplePreprocessor.inputSamples[first] * waveformYScaling });
-		auto previous = point;
-		for (int i = first + step; i < last; i += step) {
-			auto sample = *std::max_element(samplePreprocessor.inputSamples.begin() + i - step, samplePreprocessor.inputSamples.begin() + i, [](auto a, auto b) { return a * a < b * b; });
-			const auto point = transform.apply({ static_cast<double>(i), -sample * waveformYScaling });
-			painter.line(previous, point);
-			previous = point;
-		}
-	} else {
-		for (int i = first + step; i < last; i += step) {
-			auto [min, max] = std::minmax_element(samplePreprocessor.inputSamples.begin() + i - step, samplePreprocessor.inputSamples.begin() + i);
-			const auto p1 = transform.apply({ static_cast<double>(i), -*min * waveformYScaling });
-			const auto p2 = transform.apply({ static_cast<double>(i), -*max * waveformYScaling });
-			painter.line(p1, p2);
-		}
-	}
-
-	if (transform.apply({ 0, 0, 1, 0 }).width > targetSize.x / 20.) {
-		for (int i = first; i < last; ++i) {
-			const auto point = transform.apply({ static_cast<double>(i), -samplePreprocessor.inputSamples[i] * waveformYScaling });
-			painter.point(point, 5.0);
-		}
-	}
-}
-
-void table_preprocessing::drawDraggingRect(MaxPainter& painter) {
-	if (dragging && button == MouseEvent::Button::Right) {
-
-		// enable dashed stroke style
-		c74::max::t_jgraphics* g = painter.t;
-		double a[] = { 4, 4 };
-		jgraphics_set_dash(g, a, 2, 0);
-
-		Rect rect = { currentMousePoint, mouseDownPoint };
-		rect.y = margin;
-		rect.height = painter.getHeight() - 2 * margin;
-		if (rect.width > 0) {
-			painter.rectOutline(rect, 1.0);
-		}
-		// disable dashed stroke style
-		jgraphics_set_dash(g, 0, 0, 0);
-	}
-}
-
-void table_preprocessing::drawZeroCrossings(MaxPainter& painter) {
-	// get first/last visible sample
-	const int first = std::max(0., transform.fromX(0) - 1.);
-	const int last = std::min<double>(samplePreprocessor.inputSamples.size(), std::ceil(transform.fromX(targetSize.x) + 1.));
-	const int step = std::max<int>(1, (last - first) / (targetSize.x * 10.));
-
-	if (step > 5) return; // dont draw crossings when they get too dense
-	for (double value : samplePreprocessor.zeroCrossings) {
-		if (value < first) continue; // only draw visible crossings
-		if (value > last) break;
-		const auto p1 = transform.apply({ value, 1. });
-		const auto p2 = transform.apply({ value, -1. });
-		painter.line(p1, p2);
-	}
-}
-
-void table_preprocessing::drawOverlayRects(MaxPainter& painter) {
-	if (mode == Mode::free) {
-		if (freeSelection.first == freeSelection.second)
-			return;
-		const auto r = transform.apply({ { freeSelection.first, 1 }, { freeSelection.second, -1 } });
-		painter.rect(r);
-	} else if (mode == Mode::zeros) {
-		if (zerosSelection.first == zerosSelection.second)
-			return;
-		const auto r = transform.apply({ { zerosSelection.first, 1 }, { zerosSelection.second, -1 } });
-		painter.rect(r);
-	}
-}
-
-void table_preprocessing::inputSamplesChanged() {
-	resetTransform();
-}
-
-void table_preprocessing::resetTransform() {
-	dataRange = Rect::fromBounds(0, samplePreprocessor.inputSamples.size(), -1, 1);
-	transform = Transform::MapRect(dataRange, waveformView);
-}
-
-void table_preprocessing::targetResized(double width, double height) {
-	const auto currentViewRect = transform.from(waveformView);
-	targetSize = { width, height };
-	waveformView.resize(width - 2 * margin, height - 2 * margin).moveTo({ margin, margin });
-	transform = Transform::MapRect(currentViewRect, waveformView);
-	//transform = Transform::MapRect(dataRange, waveformView);
-}
-
-void table_preprocessing::constrainViewTransform() {
-	transform.ensureWithin(dataRange, waveformView);
-}
-
-
 
 void table_preprocessing::sampleDroppedImpl() {
 	inputBuffer.set(inputBufferName);
@@ -519,84 +267,6 @@ void table_preprocessing::sampleDroppedImpl() {
 void table_preprocessing::setSampleData(const std::vector<float>& data) {
 	samplePreprocessor.setSampleData(data);
 	redraw(); //show new samples
-
-	inputSamplesChanged();
-}
-
-
-void table_preprocessing::mousedownImpl(const MouseEvent& e) {
-	if (samplePreprocessor.inputSamples.empty()) return;
-	mouseDownPoint = { e.x, e.y };
-	currentMousePoint = mouseDownPoint;
-	dragging = true;
-	this->button = e.button;
-	redraw();
-}
-
-void table_preprocessing::mousedragImpl(const MouseEvent& e) {
-	if (samplePreprocessor.inputSamples.empty())
-		return;
-	Point point{ e.x, e.y };
-
-	if (e.button == MouseEvent::Button::Left) {
-		if (mode == Mode::free) {
-			updateFreeSelection(mouseDownPoint, point);
-		} else if (mode == Mode::zeros) {
-			updateZerosSelection(mouseDownPoint, point);
-		}
-	} else if (e.button == MouseEvent::Button::Middle) {
-		auto p = point - currentMousePoint;
-		transform.preTranslate(p);
-		constrainViewTransform();
-	}
-	currentMousePoint = point;
-	redraw();
-}
-
-void table_preprocessing::mouseupImpl(const MouseEvent& e) {
-	if (samplePreprocessor.inputSamples.empty()) return;
-	Point point{ e.x, e.y };
-
-	currentMousePoint = point;
-	if (e.button == MouseEvent::Button::Left) {
-		if (mode == Mode::free) {
-			updateFreeSelection(mouseDownPoint, point);
-		} else if (mode == Mode::zeros) {
-			updateZerosSelection(mouseDownPoint, point);
-		}
-	} else if (e.button == MouseEvent::Button::Right) {
-		if (dragging) {
-			Rect r{ mouseDownPoint, point };
-			r.y = waveformView.y;
-			r.height = waveformView.height;
-			if (r.width > 0) {
-				auto r2 = transform.from(r);
-				transform = Transform::MapRect(r2, waveformView);
-				constrainViewTransform();
-			}
-		}
-	}
-	dragging = false;
-	const auto now = std::chrono::system_clock::now();
-	if (now - clickTime < std::chrono::milliseconds(300)) {
-		resetTransform();
-	}
-	clickTime = now;
-	redraw();
-}
-
-void table_preprocessing::mousewheelImpl(const MouseEvent& e) {
-	const double speed = e.isControlDown() ? fastZoomSpeed : zoomSpeed;
-	const auto delta = e.deltaY > 0 ? speed : 1 / speed;
-	if (delta > 1. && transform.apply({ 0, 0, 1, 0 }).width > targetSize.x) return;
-	transform.scaleAround(e.x, e.y, delta, 1);
-	constrainViewTransform();
-	redraw();
-}
-
-double table_preprocessing::nearestZeroCrossing(double sampleIdx) const {
-	if (samplePreprocessor.zeroCrossings.empty()) { return 0.; }
-	return *std::min_element(samplePreprocessor.zeroCrossings.begin(), samplePreprocessor.zeroCrossings.end(), [sampleIdx](auto a, auto b) { return std::abs(a - sampleIdx) < std::abs(b - sampleIdx); });
 }
 
 MouseEvent::Button table_preprocessing::getButton(const event& e) const {
@@ -609,5 +279,3 @@ MouseEvent::Button table_preprocessing::getButton(const event& e) const {
 	return MouseEvent::Button::Left;
 }
 
-
-MIN_EXTERNAL(table_preprocessing);
